requirements.txt
runpod>=1.6.0
torch>=2.1.0
transformers>=4.40.0
pillow>=10.0.0
accelerate>=0.27.0
qwen-vl-utils
torchvision
